refactor: object utils
